requirements.txt
affine==2.4.0
amqp==5.3.1
asgiref==3.10.0
astroid==3.0.3
asttokens==3.0.0
attrs==25.4.0
billiard==4.2.2
black==23.11.0
boto3==1.29.7
botocore==1.32.7
celery==5.3.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cligj==0.7.2
coverage==7.11.1
cron_descriptor==2.0.6
croniter==2.0.1
cryptography==46.0.3
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
dj-database-url==2.1.0
Django==4.2.7
django-allauth==0.57.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.3
django-storages==1.14.2
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
executing==2.2.1
factory-boy==3.3.0
Faker==20.1.0
fiona==1.10.1
flake8==6.1.0
geopandas==0.14.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
ipdb==0.13.13
ipython==8.18.1
isort==5.12.0
jedi==0.19.2
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.0
matplotlib-inline==0.2.1
mccabe==0.7.0
mypy==1.7.1
mypy_extensions==1.1.0
numpy==2.3.4
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.11.1
pycparser==2.23
pyflakes==3.1.0
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.0.3
pyparsing==3.2.5
pyproj==3.6.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-crontab==3.3.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
rasterio==1.3.9
redis==5.0.1
referencing==0.37.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.28.0
s3transfer==0.8.0
sentry-sdk==1.38.0
setuptools==80.9.0
shapely==2.0.2
six==1.17.0
snuggs==1.4.7
sqlparse==0.5.3
stack-data==0.6.3
tomlkit==0.13.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.0.7
vine==5.1.0
watchdog==3.0.0
wcwidth==0.2.14
